Move repository types to the new `repository` module in aggregator
