+++ mithril-aggregator/src/database/provider/epoch_setting.rs
@@ -1,7 +1,4 @@
-use std::sync::Arc;
-
 use anyhow::Context;
-use async_trait::async_trait;
 use sqlite::Value;
 
 use mithril_common::{
@@ -11,10 +8,8 @@ use mithril_common::{
 use mithril_persistence::sqlite::{
     EntityCursor, Provider, SourceAlias, SqLiteEntity, SqliteConnection, WhereCondition,
 };
-use mithril_persistence::store::adapter::AdapterError;
 
 use crate::database::record::EpochSettingRecord;
-use crate::ProtocolParametersStorer;
 
 /// Simple queries to retrieve [EpochSettingRecord] from the sqlite database.
 pub(crate) struct EpochSettingProvider<'client> {
@@ -72,7 +67,8 @@ impl<'conn> UpdateEpochSettingProvider<'conn> {
         Self { connection }
     }
 
-    fn get_update_condition(
+    // todo: not pub
+    pub fn get_update_condition(
         &self,
         epoch: Epoch,
         protocol_parameters: ProtocolParameters,
@@ -88,7 +84,7 @@ impl<'conn> UpdateEpochSettingProvider<'conn> {
         )
     }
 
-    fn persist(
+    pub fn persist(
         &self,
         epoch: Epoch,
         protocol_parameters: ProtocolParameters,
@@ -180,68 +176,13 @@ impl<'conn> DeleteEpochSettingProvider<'conn> {
     }
 }
 
-/// Service to deal with epoch settings (read & write).
-pub struct EpochSettingStore {
-    connection: Arc<SqliteConnection>,
-
-    /// Number of epochs before previous records will be pruned at the next call to
-    /// [save_protocol_parameters][EpochSettingStore::save_protocol_parameters].
-    retention_limit: Option<u64>,
-}
-
-impl EpochSettingStore {
-    /// Create a new EpochSetting service
-    pub fn new(connection: Arc<SqliteConnection>, retention_limit: Option<u64>) -> Self {
-        Self {
-            connection,
-            retention_limit,
-        }
-    }
-}
-
-#[async_trait]
-impl ProtocolParametersStorer for EpochSettingStore {
-    async fn save_protocol_parameters(
-        &self,
-        epoch: Epoch,
-        protocol_parameters: ProtocolParameters,
-    ) -> StdResult<Option<ProtocolParameters>> {
-        let provider = UpdateEpochSettingProvider::new(&self.connection);
-        let epoch_setting_record = provider.persist(epoch, protocol_parameters).map_err(|e| {
-            AdapterError::GeneralError(e.context("persist protocol parameters failure"))
-        })?;
-
-        // Prune useless old epoch settings.
-        if let Some(threshold) = self.retention_limit {
-            let _ = DeleteEpochSettingProvider::new(&self.connection)
-                .prune(epoch - threshold)
-                .map_err(AdapterError::QueryError)?
-                .count();
-        }
-
-        Ok(Some(epoch_setting_record.protocol_parameters))
-    }
-
-    async fn get_protocol_parameters(&self, epoch: Epoch) -> StdResult<Option<ProtocolParameters>> {
-        let provider = EpochSettingProvider::new(&self.connection);
-        let mut cursor = provider
-            .get_by_epoch(&epoch)
-            .map_err(|e| AdapterError::GeneralError(e.context("Could not get epoch setting")))?;
-
-        if let Some(epoch_setting_record) = cursor.next() {
-            return Ok(Some(epoch_setting_record.protocol_parameters));
-        }
-        Ok(None)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use sqlite::Connection;
 
     use mithril_common::test_utils::fake_data;
 
-    use crate::database::test_helper::apply_all_migrations_to_db;
+    use crate::database::test_helper::{apply_all_migrations_to_db, insert_epoch_settings};
 
     use super::*;
 
@@ -250,36 +191,7 @@ mod tests {
         epoch_to_insert_settings: &[u64],
     ) -> StdResult<()> {
         apply_all_migrations_to_db(connection)?;
-
-        let query = {
-            // leverage the expanded parameter from this provider which is unit
-            // tested on its own above.
-            let update_provider = UpdateEpochSettingProvider::new(connection);
-            let (sql_values, _) = update_provider
-                .get_update_condition(Epoch(1), ProtocolParameters::new(1, 2, 1.0))
-                .expand();
-
-            format!("insert into epoch_setting {sql_values}")
-        };
-
-        for (epoch, protocol_parameters) in epoch_to_insert_settings
-            .iter()
-            .map(|epoch| (epoch, ProtocolParameters::new(*epoch, epoch + 1, 1.0)))
-        {
-            let mut statement = connection.prepare(&query)?;
-            statement
-                .bind::<&[(_, Value)]>(&[
-                    (1, Value::Integer(*epoch as i64)),
-                    (
-                        2,
-                        serde_json::to_string(&protocol_parameters).unwrap().into(),
-                    ),
-                ])
-                .unwrap();
-
-            statement.next().unwrap();
-        }
-
+        insert_epoch_settings(connection, epoch_to_insert_settings)?;
         Ok(())
     }
 
@@ -446,35 +358,5 @@ mod tests {
         let cursor = provider.get_by_epoch(&Epoch(2)).unwrap();
 
         assert_eq!(1, cursor.count());
-    }
-
-    #[tokio::test]
-    async fn save_protocol_parameters_prune_older_epoch_settings() {
-        let connection = Connection::open_thread_safe(":memory:").unwrap();
-        const EPOCH_SETTING_PRUNE_EPOCH_THRESHOLD: u64 = 5;
-        setup_epoch_setting_db(&connection, &[1, 2]).unwrap();
-        let store = EpochSettingStore::new(
-            Arc::new(connection),
-            Some(EPOCH_SETTING_PRUNE_EPOCH_THRESHOLD),
-        );
-
-        store
-            .save_protocol_parameters(
-                Epoch(2) + EPOCH_SETTING_PRUNE_EPOCH_THRESHOLD,
-                fake_data::protocol_parameters(),
-            )
-            .await
-            .expect("saving protocol parameters should not fails");
-        let epoch1_params = store.get_protocol_parameters(Epoch(1)).await.unwrap();
-        let epoch2_params = store.get_protocol_parameters(Epoch(2)).await.unwrap();
-
-        assert!(
-            epoch1_params.is_none(),
-            "Protocol parameters at epoch 1 should have been pruned",
-        );
-        assert!(
-            epoch2_params.is_some(),
-            "Protocol parameters at epoch 2 should still exist",
-        );
     }
 }
