+++ mithril-aggregator/src/database/provider/mod.rs
@@ -9,12 +9,12 @@ mod signer_registration;
 mod single_signature;
 mod stake_pool;
 
-pub use cardano_transaction::*;
-pub use certificate::*;
-pub use epoch_setting::*;
-pub use open_message::*;
-pub use signed_entity::*;
-pub use signer::*;
-pub use signer_registration::*;
-pub use single_signature::*;
-pub use stake_pool::*;
+pub(crate) use cardano_transaction::*;
+pub(crate) use certificate::*;
+pub(crate) use epoch_setting::*;
+pub(crate) use open_message::*;
+pub(crate) use signed_entity::*;
+pub(crate) use signer::*;
+pub(crate) use signer_registration::*;
+pub(crate) use single_signature::*;
+pub(crate) use stake_pool::*;
