+++ mithril-aggregator/src/database/provider/signed_entity.rs
@@ -1,20 +1,12 @@
-use std::sync::Arc;
-
-use anyhow::Context;
-use async_trait::async_trait;
 use sqlite::Value;
 
 use mithril_common::entities::SignedEntityTypeDiscriminants;
 use mithril_common::StdResult;
-use mithril_persistence::{
-    sqlite::{EntityCursor, Provider, SourceAlias, SqLiteEntity, SqliteConnection, WhereCondition},
-    store::adapter::AdapterError,
+use mithril_persistence::sqlite::{
+    EntityCursor, Provider, SourceAlias, SqLiteEntity, SqliteConnection, WhereCondition,
 };
 
 use crate::database::record::SignedEntityRecord;
-
-#[cfg(test)]
-use mockall::automock;
 
 /// Simple queries to retrieve [SignedEntityRecord] from the sqlite database.
 pub(crate) struct SignedEntityRecordProvider<'client> {
@@ -143,7 +135,10 @@ impl<'conn> InsertSignedEntityRecordProvider<'conn> {
         Self { connection }
     }
 
-    fn get_insert_condition(&self, signed_entity_record: SignedEntityRecord) -> WhereCondition {
+    pub(crate) fn get_insert_condition(
+        &self,
+        signed_entity_record: SignedEntityRecord,
+    ) -> WhereCondition {
         WhereCondition::new(
             "(signed_entity_id, signed_entity_type_id, certificate_id, beacon, artifact, created_at) values (?*, ?*, ?*, ?*, ?*, ?*)",
             vec![
@@ -157,7 +152,10 @@ impl<'conn> InsertSignedEntityRecordProvider<'conn> {
         )
     }
 
-    fn persist(&self, signed_entity_record: SignedEntityRecord) -> StdResult<SignedEntityRecord> {
+    pub(crate) fn persist(
+        &self,
+        signed_entity_record: SignedEntityRecord,
+    ) -> StdResult<SignedEntityRecord> {
         let filters = self.get_insert_condition(signed_entity_record.clone());
 
         let entity = self.find(filters)?.next().unwrap_or_else(|| {
@@ -218,7 +216,10 @@ where signed_entity_id = ?*";
         Ok(WhereCondition::new(expression, parameters))
     }
 
-    fn persist(&self, signed_entity_record: &SignedEntityRecord) -> StdResult<SignedEntityRecord> {
+    pub(crate) fn persist(
+        &self,
+        signed_entity_record: &SignedEntityRecord,
+    ) -> StdResult<SignedEntityRecord> {
         let filters = self.get_update_condition(signed_entity_record)?;
         let mut cursor = self.find(filters)?;
 
@@ -246,140 +247,6 @@ impl<'client> Provider<'client> for UpdateSignedEntityProvider<'client> {
     }
 }
 
-/// Signed entity storer trait
-#[cfg_attr(test, automock)]
-#[async_trait]
-pub trait SignedEntityStorer: Sync + Send {
-    /// Store a signed entity
-    async fn store_signed_entity(&self, signed_entity: &SignedEntityRecord) -> StdResult<()>;
-
-    /// Get signed entity type
-    async fn get_signed_entity(
-        &self,
-        signed_entity_id: &str,
-    ) -> StdResult<Option<SignedEntityRecord>>;
-
-    /// Get signed entity type by certificate id
-    async fn get_signed_entity_by_certificate_id(
-        &self,
-        certificate_hash: &str,
-    ) -> StdResult<Option<SignedEntityRecord>>;
-
-    /// Get signed entities type by certificates ids
-    async fn get_signed_entities_by_certificates_ids<'a>(
-        &self,
-        certificates_ids: &[&'a str],
-    ) -> StdResult<Vec<SignedEntityRecord>>;
-
-    /// Get last signed entities by signed entity type
-    async fn get_last_signed_entities_by_type(
-        &self,
-        signed_entity_type_id: &SignedEntityTypeDiscriminants,
-        total: usize,
-    ) -> StdResult<Vec<SignedEntityRecord>>;
-
-    /// Perform an update for all the given signed entities.
-    async fn update_signed_entities(
-        &self,
-        signed_entities: Vec<SignedEntityRecord>,
-    ) -> StdResult<Vec<SignedEntityRecord>>;
-}
-
-/// Service to deal with signed_entity (read & write).
-pub struct SignedEntityStoreAdapter {
-    connection: Arc<SqliteConnection>,
-}
-
-impl SignedEntityStoreAdapter {
-    /// Create a new SignedEntityStoreAdapter service
-    pub fn new(connection: Arc<SqliteConnection>) -> Self {
-        Self { connection }
-    }
-}
-
-#[async_trait]
-impl SignedEntityStorer for SignedEntityStoreAdapter {
-    async fn store_signed_entity(&self, signed_entity: &SignedEntityRecord) -> StdResult<()> {
-        let provider = InsertSignedEntityRecordProvider::new(&self.connection);
-        let _signed_entity_record = provider.persist(signed_entity.to_owned())?;
-
-        Ok(())
-    }
-
-    async fn get_signed_entity(
-        &self,
-        signed_entity_id: &str,
-    ) -> StdResult<Option<SignedEntityRecord>> {
-        let provider = SignedEntityRecordProvider::new(&self.connection);
-        let mut cursor = provider
-            .get_by_signed_entity_id(signed_entity_id)
-            .with_context(|| format!("get signed entity by id failure, id: {signed_entity_id}"))
-            .map_err(AdapterError::GeneralError)?;
-        let signed_entity = cursor.next();
-
-        Ok(signed_entity)
-    }
-
-    async fn get_signed_entity_by_certificate_id(
-        &self,
-        certificate_id: &str,
-    ) -> StdResult<Option<SignedEntityRecord>> {
-        let provider = SignedEntityRecordProvider::new(&self.connection);
-        let mut cursor = provider
-            .get_by_certificate_id(certificate_id)
-            .with_context(|| {
-                format!(
-                    "get signed entity by certificate id failure, certificate_id: {certificate_id}"
-                )
-            })
-            .map_err(AdapterError::GeneralError)?;
-        let signed_entity = cursor.next();
-
-        Ok(signed_entity)
-    }
-
-    async fn get_signed_entities_by_certificates_ids<'a>(
-        &self,
-        certificates_ids: &[&'a str],
-    ) -> StdResult<Vec<SignedEntityRecord>> {
-        let provider = SignedEntityRecordProvider::new(&self.connection);
-        let cursor = provider.get_by_certificates_ids(certificates_ids)?;
-
-        Ok(cursor.collect())
-    }
-
-    async fn get_last_signed_entities_by_type(
-        &self,
-        signed_entity_type_id: &SignedEntityTypeDiscriminants,
-        total: usize,
-    ) -> StdResult<Vec<SignedEntityRecord>> {
-        let provider = SignedEntityRecordProvider::new(&self.connection);
-        let cursor = provider
-            .get_by_signed_entity_type(signed_entity_type_id)
-            .with_context(|| {
-                format!("get last signed entity by type failure, type: {signed_entity_type_id:?}")
-            })
-            .map_err(AdapterError::GeneralError)?;
-        let signed_entities: Vec<SignedEntityRecord> = cursor.take(total).collect();
-
-        Ok(signed_entities)
-    }
-
-    async fn update_signed_entities(
-        &self,
-        signed_entities: Vec<SignedEntityRecord>,
-    ) -> StdResult<Vec<SignedEntityRecord>> {
-        let provider = UpdateSignedEntityProvider::new(&self.connection);
-        let mut updated_records = vec![];
-
-        for record in signed_entities {
-            updated_records.push(provider.persist(&record)?);
-        }
-
-        Ok(updated_records)
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use chrono::{DateTime, Utc};
@@ -390,30 +257,11 @@ mod tests {
     };
     use mithril_common::{entities::CardanoDbBeacon, test_utils::fake_data};
 
-    use crate::database::test_helper::{apply_all_migrations_to_db, disable_foreign_key_support};
+    use crate::database::test_helper::{
+        apply_all_migrations_to_db, disable_foreign_key_support, insert_signed_entities,
+    };
 
     use super::*;
-
-    pub fn fake_signed_entity_records(total_records: usize) -> Vec<SignedEntityRecord> {
-        let snapshots = fake_data::snapshots(total_records as u64);
-        (0..total_records)
-            .map(|idx| {
-                let snapshot = snapshots.get(idx).unwrap().to_owned();
-                let entity = serde_json::to_string(&snapshot).unwrap();
-                SignedEntityRecord {
-                    signed_entity_id: snapshot.digest,
-                    signed_entity_type: SignedEntityType::CardanoImmutableFilesFull(
-                        snapshot.beacon,
-                    ),
-                    certificate_id: format!("certificate-{idx}"),
-                    artifact: entity,
-                    created_at: DateTime::parse_from_rfc3339("2023-01-19T13:43:05.618857482Z")
-                        .unwrap()
-                        .with_timezone(&Utc),
-                }
-            })
-            .collect()
-    }
 
     pub fn setup_signed_entity_db(
         connection: &SqliteConnection,
@@ -421,48 +269,7 @@ mod tests {
     ) -> StdResult<()> {
         apply_all_migrations_to_db(connection)?;
         disable_foreign_key_support(connection)?;
-
-        if signed_entity_records.is_empty() {
-            return Ok(());
-        }
-
-        let query = {
-            // leverage the expanded parameter from this provider which is unit
-            // tested on its own above.
-            let insert_provider = InsertSignedEntityRecordProvider::new(connection);
-            let (sql_values, _) = insert_provider
-                .get_insert_condition(signed_entity_records.first().unwrap().to_owned())
-                .expand();
-            format!("insert into signed_entity {sql_values}")
-        };
-
-        for signed_entity_record in signed_entity_records {
-            let mut statement = connection.prepare(&query)?;
-            statement
-                .bind::<&[(_, Value)]>(&[
-                    (1, signed_entity_record.signed_entity_id.into()),
-                    (
-                        2,
-                        i64::try_from(signed_entity_record.signed_entity_type.index())
-                            .unwrap()
-                            .into(),
-                    ),
-                    (3, signed_entity_record.certificate_id.into()),
-                    (
-                        4,
-                        signed_entity_record
-                            .signed_entity_type
-                            .get_json_beacon()
-                            .unwrap()
-                            .into(),
-                    ),
-                    (5, signed_entity_record.artifact.into()),
-                    (6, signed_entity_record.created_at.to_rfc3339().into()),
-                ])
-                .unwrap();
-
-            statement.next().unwrap();
-        }
+        insert_signed_entities(connection, signed_entity_records)?;
 
         Ok(())
     }
@@ -648,7 +455,7 @@ mod tests {
 
     #[test]
     fn test_get_signed_entity_records() {
-        let signed_entity_records = fake_signed_entity_records(5);
+        let signed_entity_records = SignedEntityRecord::fake_records(5);
 
         let connection = Connection::open_thread_safe(":memory:").unwrap();
         setup_signed_entity_db(&connection, signed_entity_records.clone()).unwrap();
@@ -683,47 +490,9 @@ mod tests {
         assert_eq!(expected_signed_entity_records, signed_entity_records);
     }
 
-    #[tokio::test]
-    async fn test_get_signed_entity_record_by_certificate_id() {
-        let expected_record = fake_signed_entity_records(1).remove(0);
-        let connection = Connection::open_thread_safe(":memory:").unwrap();
-        setup_signed_entity_db(&connection, vec![expected_record.clone()]).unwrap();
-        let store = SignedEntityStoreAdapter::new(Arc::new(connection));
-
-        let record = store
-            .get_signed_entity_by_certificate_id(&expected_record.certificate_id)
-            .await
-            .expect("querying signed entity record by certificate id should not fail");
-
-        assert_eq!(Some(expected_record), record);
-    }
-
-    #[tokio::test]
-    async fn test_get_signed_entity_record_by_certificates_ids() {
-        let expected_records = fake_signed_entity_records(3);
-        let connection = Connection::open_thread_safe(":memory:").unwrap();
-        setup_signed_entity_db(&connection, expected_records.clone()).unwrap();
-        let store = SignedEntityStoreAdapter::new(Arc::new(connection));
-        let certificates_ids: Vec<&str> = expected_records
-            .iter()
-            .map(|r| r.certificate_id.as_str())
-            .collect();
-
-        let queried_records = store
-            .get_signed_entities_by_certificates_ids(&certificates_ids)
-            .await
-            .expect("querying signed entity record by certificates ids should not fail");
-
-        assert_eq!(
-            // Records are inserted older to earlier and queried the other way round
-            expected_records.into_iter().rev().collect::<Vec<_>>(),
-            queried_records
-        );
-    }
-
     #[test]
     fn test_insert_signed_entity_record() {
-        let signed_entity_records = fake_signed_entity_records(5);
+        let signed_entity_records = SignedEntityRecord::fake_records(5);
 
         let connection = Connection::open_thread_safe(":memory:").unwrap();
         setup_signed_entity_db(&connection, Vec::new()).unwrap();
@@ -776,43 +545,5 @@ mod tests {
             ],
             params
         );
-    }
-
-    #[tokio::test]
-    async fn update_only_given_entities() {
-        let mut signed_entity_records = fake_signed_entity_records(5);
-
-        let connection = Connection::open_thread_safe(":memory:").unwrap();
-        setup_signed_entity_db(&connection, signed_entity_records.clone()).unwrap();
-        let store = SignedEntityStoreAdapter::new(Arc::new(connection));
-
-        let records_to_update: Vec<SignedEntityRecord> = signed_entity_records
-            .drain(2..)
-            .map(|mut r| {
-                r.certificate_id = format!("updated-{}", r.certificate_id);
-                r
-            })
-            .collect();
-        let expected_records: Vec<SignedEntityRecord> = signed_entity_records
-            .into_iter()
-            .chain(records_to_update.clone())
-            .rev() // Records are returned from latest to oldest
-            .collect();
-
-        let updated_records = store
-            .update_signed_entities(records_to_update.clone())
-            .await
-            .expect("updating signed entities should not fail");
-
-        let stored_records = store
-            .get_last_signed_entities_by_type(
-                &SignedEntityTypeDiscriminants::CardanoImmutableFilesFull,
-                usize::MAX,
-            )
-            .await
-            .expect("getting signed entities should not fail");
-
-        assert_eq!(records_to_update, updated_records);
-        assert_eq!(expected_records, stored_records);
     }
 }
