+++ mithril-aggregator/src/database/provider/signer.rs
@@ -1,20 +1,13 @@
-use std::collections::HashMap;
 use std::iter::repeat;
-use std::sync::Arc;
 
-use async_trait::async_trait;
-use chrono::Utc;
+use sqlite::Value;
+
 use mithril_common::StdResult;
 use mithril_persistence::sqlite::{
     EntityCursor, Provider, SourceAlias, SqLiteEntity, SqliteConnection, WhereCondition,
 };
-use sqlite::Value;
 
 use crate::database::record::SignerRecord;
-use crate::signer_registerer::SignerRecorder;
-
-#[cfg(test)]
-use mockall::automock;
 
 /// Simple queries to retrieve [SignerRecord] from the sqlite database.
 pub(crate) struct SignerRecordProvider<'client> {
@@ -99,7 +92,7 @@ impl<'conn> RegisterSignerRecordProvider<'conn> {
         )
     }
 
-    fn persist(&self, signer_record: SignerRecord) -> StdResult<SignerRecord> {
+    pub(crate) fn persist(&self, signer_record: SignerRecord) -> StdResult<SignerRecord> {
         let filters = self.get_register_condition(signer_record.clone());
 
         let entity = self.find(filters)?.next().unwrap_or_else(|| {
@@ -144,7 +137,7 @@ impl<'conn> ImportSignerRecordProvider<'conn> {
         Self { connection }
     }
 
-    fn get_import_condition(&self, signer_records: Vec<SignerRecord>) -> WhereCondition {
+    pub(crate) fn get_import_condition(&self, signer_records: Vec<SignerRecord>) -> WhereCondition {
         let columns = "(signer_id, pool_ticker, created_at, updated_at, last_registered_at)";
         let values_columns: Vec<&str> = repeat("(?*, ?*, ?*, ?*, ?*)")
             .take(signer_records.len())
@@ -174,7 +167,7 @@ impl<'conn> ImportSignerRecordProvider<'conn> {
         )
     }
 
-    fn persist(&self, signer_record: SignerRecord) -> StdResult<SignerRecord> {
+    pub(crate) fn persist(&self, signer_record: SignerRecord) -> StdResult<SignerRecord> {
         let filters = self.get_import_condition(vec![signer_record.clone()]);
 
         let entity = self.find(filters)?.next().unwrap_or_else(|| {
@@ -184,7 +177,10 @@ impl<'conn> ImportSignerRecordProvider<'conn> {
         Ok(entity)
     }
 
-    fn persist_many(&self, signer_records: Vec<SignerRecord>) -> StdResult<Vec<SignerRecord>> {
+    pub(crate) fn persist_many(
+        &self,
+        signer_records: Vec<SignerRecord>,
+    ) -> StdResult<Vec<SignerRecord>> {
         let filters = self.get_import_condition(signer_records);
 
         Ok(self.find(filters)?.collect())
@@ -211,182 +207,23 @@ impl<'conn> Provider<'conn> for ImportSignerRecordProvider<'conn> {
     }
 }
 
-/// Service to get [SignerRecord].
-#[cfg_attr(test, automock)]
-#[async_trait]
-pub trait SignerGetter: Sync + Send {
-    /// Return all stored records.
-    async fn get_all(&self) -> StdResult<Vec<SignerRecord>>;
-}
-
-/// Service to deal with signer (read & write).
-pub struct SignerStore {
-    connection: Arc<SqliteConnection>,
-}
-
-impl SignerStore {
-    /// Create a new SignerStore service
-    pub fn new(connection: Arc<SqliteConnection>) -> Self {
-        Self { connection }
-    }
-
-    /// Import a signer in the database, its last_registered_at date will be left empty
-    pub async fn import_signer(
-        &self,
-        signer_id: String,
-        pool_ticker: Option<String>,
-    ) -> StdResult<()> {
-        let provider = ImportSignerRecordProvider::new(&self.connection);
-        let created_at = Utc::now();
-        let updated_at = created_at;
-        let signer_record = SignerRecord {
-            signer_id,
-            pool_ticker,
-            created_at,
-            updated_at,
-            last_registered_at: None,
-        };
-        provider.persist(signer_record)?;
-
-        Ok(())
-    }
-
-    /// Create many signers at once in the database, their last_registered_at date will be left empty
-    pub async fn import_many_signers(
-        &self,
-        pool_ticker_by_id: HashMap<String, Option<String>>,
-    ) -> StdResult<()> {
-        let provider = ImportSignerRecordProvider::new(&self.connection);
-
-        let created_at = Utc::now();
-        let updated_at = created_at;
-        let signer_records: Vec<_> = pool_ticker_by_id
-            .into_iter()
-            .map(|(signer_id, pool_ticker)| SignerRecord {
-                signer_id,
-                pool_ticker,
-                created_at,
-                updated_at,
-                last_registered_at: None,
-            })
-            .collect();
-
-        provider.persist_many(signer_records)?;
-
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl SignerRecorder for SignerStore {
-    async fn record_signer_registration(&self, signer_id: String) -> StdResult<()> {
-        let provider = RegisterSignerRecordProvider::new(&self.connection);
-        let created_at = Utc::now();
-        let updated_at = created_at;
-        let registered_at = Some(created_at);
-        let signer_record = SignerRecord {
-            signer_id,
-            pool_ticker: None,
-            created_at,
-            updated_at,
-            last_registered_at: registered_at,
-        };
-        provider.persist(signer_record)?;
-
-        Ok(())
-    }
-}
-
-#[async_trait]
-impl SignerGetter for SignerStore {
-    async fn get_all(&self) -> StdResult<Vec<SignerRecord>> {
-        let provider = SignerRecordProvider::new(&self.connection);
-        let cursor = provider.get_all()?;
-
-        Ok(cursor.collect())
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use std::collections::BTreeMap;
-
-    use chrono::{DateTime, Duration};
+    use chrono::Duration;
     use sqlite::Connection;
 
     use mithril_common::StdResult;
 
-    use crate::database::test_helper::apply_all_migrations_to_db;
+    use crate::database::test_helper::{apply_all_migrations_to_db, insert_signers};
 
     use super::*;
 
-    fn fake_signer_records(total_records: usize) -> Vec<SignerRecord> {
-        (0..total_records)
-            .map(|idx| SignerRecord {
-                signer_id: format!("signer-{idx}"),
-                pool_ticker: Some(format!("pool-ticker-{idx}")),
-                created_at: DateTime::parse_from_rfc3339("2023-01-19T13:43:05.618857482Z")
-                    .unwrap()
-                    .with_timezone(&Utc),
-                updated_at: DateTime::parse_from_rfc3339("2024-01-19T13:43:05.618857482Z")
-                    .unwrap()
-                    .with_timezone(&Utc),
-                last_registered_at: Some(
-                    DateTime::parse_from_rfc3339("2023-01-19T13:43:05.618857482Z")
-                        .unwrap()
-                        .with_timezone(&Utc),
-                ),
-            })
-            .collect()
-    }
-
     pub fn setup_signer_db(
         connection: &SqliteConnection,
-        signer_records: Vec<SignerRecord>,
+        records: Vec<SignerRecord>,
     ) -> StdResult<()> {
         apply_all_migrations_to_db(connection)?;
-
-        if signer_records.is_empty() {
-            return Ok(());
-        }
-
-        let query = {
-            // leverage the expanded parameter from this provider which is unit
-            // tested on its own above.
-            let update_provider = ImportSignerRecordProvider::new(connection);
-            let (sql_values, _) = update_provider
-                .get_import_condition(vec![signer_records.first().unwrap().to_owned()])
-                .expand();
-            format!("insert into signer {sql_values}")
-        };
-
-        for signer_record in signer_records {
-            let mut statement = connection.prepare(&query)?;
-            statement
-                .bind::<&[(_, Value)]>(&[
-                    (1, signer_record.signer_id.into()),
-                    (
-                        2,
-                        signer_record
-                            .pool_ticker
-                            .map(Value::String)
-                            .unwrap_or(Value::Null),
-                    ),
-                    (3, signer_record.created_at.to_rfc3339().into()),
-                    (4, signer_record.updated_at.to_rfc3339().into()),
-                    (
-                        5,
-                        signer_record
-                            .last_registered_at
-                            .map(|d| Value::String(d.to_rfc3339()))
-                            .unwrap_or(Value::Null),
-                    ),
-                ])
-                .unwrap();
-            statement.next().unwrap();
-        }
-
-        Ok(())
+        insert_signers(connection, records)
     }
 
     #[test]
@@ -417,7 +254,7 @@ mod tests {
 
     #[test]
     fn insert_signer_record() {
-        let signer_record = fake_signer_records(1).first().unwrap().to_owned();
+        let signer_record = SignerRecord::fake_records(1).first().unwrap().to_owned();
         let connection = Connection::open_thread_safe(":memory:").unwrap();
         let provider = RegisterSignerRecordProvider::new(&connection);
         let condition = provider.get_register_condition(signer_record.clone());
@@ -441,7 +278,7 @@ mod tests {
 
     #[test]
     fn update_signer_record() {
-        let signer_records = fake_signer_records(2);
+        let signer_records = SignerRecord::fake_records(2);
         let connection = Connection::open_thread_safe(":memory:").unwrap();
         let provider = ImportSignerRecordProvider::new(&connection);
         let condition = provider.get_import_condition(signer_records.clone());
@@ -470,7 +307,7 @@ mod tests {
 
     #[test]
     fn test_get_signer_records() {
-        let signer_records_fake = fake_signer_records(5);
+        let signer_records_fake = SignerRecord::fake_records(5);
 
         let connection = Connection::open_thread_safe(":memory:").unwrap();
         setup_signer_db(&connection, signer_records_fake.clone()).unwrap();
@@ -504,7 +341,7 @@ mod tests {
 
     #[test]
     fn test_insert_signer_record() {
-        let signer_records_fake = fake_signer_records(5);
+        let signer_records_fake = SignerRecord::fake_records(5);
 
         let connection = Connection::open_thread_safe(":memory:").unwrap();
         setup_signer_db(&connection, Vec::new()).unwrap();
@@ -525,7 +362,7 @@ mod tests {
 
     #[test]
     fn test_update_signer_record() {
-        let signer_records_fake = fake_signer_records(5);
+        let signer_records_fake = SignerRecord::fake_records(5);
 
         let connection = Connection::open_thread_safe(":memory:").unwrap();
         setup_signer_db(&connection, signer_records_fake.clone()).unwrap();
@@ -547,7 +384,7 @@ mod tests {
 
     #[test]
     fn test_update_many_signer_records() {
-        let mut signer_records_fake = fake_signer_records(5);
+        let mut signer_records_fake = SignerRecord::fake_records(5);
         signer_records_fake.sort_by(|a, b| a.signer_id.cmp(&b.signer_id));
 
         let connection = Connection::open_thread_safe(":memory:").unwrap();
@@ -565,119 +402,5 @@ mod tests {
         let mut saved_records = provider.persist_many(signer_records_fake.clone()).unwrap();
         saved_records.sort_by(|a, b| a.signer_id.cmp(&b.signer_id));
         assert_eq!(signer_records_fake, saved_records);
-    }
-
-    #[tokio::test]
-    async fn test_get_all_signers() {
-        let signer_records = fake_signer_records(5);
-        let expected: Vec<_> = signer_records.iter().rev().cloned().collect();
-        let connection = Connection::open_thread_safe(":memory:").unwrap();
-        setup_signer_db(&connection, signer_records).unwrap();
-
-        let store = SignerStore::new(Arc::new(connection));
-
-        let stored_signers = store
-            .get_all()
-            .await
-            .expect("getting all signers should not fail");
-
-        assert_eq!(expected, stored_signers);
-    }
-
-    #[tokio::test]
-    async fn test_signer_recorder() {
-        let signer_records_fake = fake_signer_records(5);
-
-        let connection = Connection::open_thread_safe(":memory:").unwrap();
-        setup_signer_db(&connection, Vec::new()).unwrap();
-
-        let connection = Arc::new(connection);
-        let store_recorder = SignerStore::new(connection.clone());
-
-        for signer_record in signer_records_fake.clone() {
-            store_recorder
-                .record_signer_registration(signer_record.signer_id.clone())
-                .await
-                .expect("record_signer_registration should not fail");
-            let provider = SignerRecordProvider::new(&connection);
-            let signer_records_stored: Vec<SignerRecord> = provider
-                .get_by_signer_id(signer_record.signer_id)
-                .unwrap()
-                .collect::<Vec<_>>();
-            assert_eq!(1, signer_records_stored.len());
-            assert!(
-                signer_records_stored
-                    .iter()
-                    .all(|s| s.last_registered_at.is_some()),
-                "registering a signer should set the registration date"
-            )
-        }
-    }
-
-    #[tokio::test]
-    async fn test_store_import_signer() {
-        let signer_records_fake = fake_signer_records(5);
-
-        let connection = Connection::open_thread_safe(":memory:").unwrap();
-        setup_signer_db(&connection, Vec::new()).unwrap();
-
-        let connection = Arc::new(connection);
-        let store = SignerStore::new(connection.clone());
-
-        for signer_record in signer_records_fake {
-            store
-                .import_signer(
-                    signer_record.signer_id.clone(),
-                    signer_record.pool_ticker.clone(),
-                )
-                .await
-                .expect("import_signer should not fail");
-            let provider = SignerRecordProvider::new(&connection);
-            let signer_records_stored: Vec<SignerRecord> = provider
-                .get_by_signer_id(signer_record.signer_id)
-                .unwrap()
-                .collect::<Vec<_>>();
-            assert_eq!(
-                signer_record.pool_ticker,
-                signer_records_stored[0].to_owned().pool_ticker
-            );
-            assert!(
-                signer_records_stored
-                    .iter()
-                    .all(|s| s.last_registered_at.is_none()),
-                "imported signer should not have a registration date"
-            )
-        }
-    }
-
-    #[tokio::test]
-    async fn test_store_import_many_signers() {
-        let signers_fake: BTreeMap<_, _> = fake_signer_records(5)
-            .into_iter()
-            .map(|r| (r.signer_id, r.pool_ticker))
-            .collect();
-
-        let connection = Connection::open_thread_safe(":memory:").unwrap();
-        setup_signer_db(&connection, Vec::new()).unwrap();
-        let store = SignerStore::new(Arc::new(connection));
-
-        store
-            .import_many_signers(signers_fake.clone().into_iter().collect())
-            .await
-            .expect("import_many_signers should not fail");
-
-        let signer_records_stored = store.get_all().await.unwrap();
-        let signers_stored = signer_records_stored
-            .iter()
-            .cloned()
-            .map(|r| (r.signer_id, r.pool_ticker))
-            .collect();
-        assert_eq!(signers_fake, signers_stored);
-        assert!(
-            signer_records_stored
-                .iter()
-                .all(|s| s.last_registered_at.is_none()),
-            "imported signer should not have a registration date"
-        );
     }
 }
