+++ mithril-aggregator/src/services/cardano_transactions_importer.rs
@@ -113,7 +113,7 @@ impl TransactionsImporter for CardanoTransactionsImporter {
 
 #[cfg(test)]
 mod tests {
-    use crate::database::provider::CardanoTransactionRepository;
+    use crate::database::repository::CardanoTransactionRepository;
     use crate::database::test_helper::apply_all_transactions_db_migrations;
     use mockall::mock;
     use mockall::predicate::eq;
