+++ mithril-aggregator/src/tools/certificates_hash_migrator.rs
@@ -5,7 +5,7 @@ use slog_scope::{debug, info, trace};
 
 use mithril_common::{entities::Certificate, StdResult};
 
-use crate::database::provider::{CertificateRepository, SignedEntityStorer};
+use crate::database::repository::{CertificateRepository, SignedEntityStorer};
 
 /// Tools to recompute all the certificates hashes in a aggregator database.
 pub struct CertificatesHashMigrator {
@@ -190,26 +190,19 @@ impl CertificatesHashMigrator {
 
 #[cfg(test)]
 mod test {
-    use std::{collections::HashMap, sync::Arc};
-
-    use anyhow::Context;
     use sqlite::Connection;
 
     use mithril_common::entities::{
-        Certificate, ImmutableFileNumber, SignedEntityType, SignedEntityTypeDiscriminants as Type,
-        TimePoint,
+        ImmutableFileNumber, SignedEntityType, SignedEntityTypeDiscriminants as Type, TimePoint,
     };
     use mithril_common::test_utils::fake_data;
-    use mithril_common::StdResult;
     use mithril_persistence::sqlite::SqliteConnection;
 
-    use crate::database::provider::{
-        CertificateRepository, SignedEntityStoreAdapter, SignedEntityStorer,
-    };
     use crate::database::record::{CertificateRecord, SignedEntityRecord};
+    use crate::database::repository::SignedEntityStore;
     use crate::database::test_helper::{apply_all_migrations_to_db, disable_foreign_key_support};
 
-    use super::CertificatesHashMigrator;
+    use super::*;
 
     fn connection_with_foreign_key_support() -> SqliteConnection {
         let connection = Connection::open_thread_safe(":memory:").unwrap();
@@ -304,7 +297,7 @@ mod test {
     ) -> StdResult<Vec<(Certificate, Option<SignedEntityRecord>)>> {
         let certificate_repository: CertificateRepository =
             CertificateRepository::new(connection.clone());
-        let signed_entity_store = SignedEntityStoreAdapter::new(connection.clone());
+        let signed_entity_store = SignedEntityStore::new(connection.clone());
         let mut result = vec![];
 
         for certificate in certificates.iter().cloned() {
@@ -414,7 +407,7 @@ mod test {
         let mut result = vec![];
         let certificate_repository: CertificateRepository =
             CertificateRepository::new(connection.clone());
-        let signed_entity_store = SignedEntityStoreAdapter::new(connection.clone());
+        let signed_entity_store = SignedEntityStore::new(connection.clone());
 
         let certificates = certificate_repository
             .get_latest_certificates::<Certificate>(usize::MAX)
@@ -455,7 +448,7 @@ mod test {
         // Act
         let migrator = CertificatesHashMigrator::new(
             CertificateRepository::new(sqlite_connection.clone()),
-            Arc::new(SignedEntityStoreAdapter::new(sqlite_connection.clone())),
+            Arc::new(SignedEntityStore::new(sqlite_connection.clone())),
         );
         migrator
             .migrate()
@@ -619,7 +612,7 @@ mod test {
 
         let migrator = CertificatesHashMigrator::new(
             CertificateRepository::new(connection.clone()),
-            Arc::new(SignedEntityStoreAdapter::new(connection.clone())),
+            Arc::new(SignedEntityStore::new(connection.clone())),
         );
         migrator
             .migrate()
